# save completion caches in Parse.Move.Interpret