```sh
# before
❯ hyperfine "neut clean && neut build test"
Benchmark 1: neut clean && neut build test
  Time (mean ± σ):     598.6 ms ±   9.2 ms    [User: 1427.3 ms, System: 560.8 ms]
  Range (min … max):   585.9 ms … 614.1 ms    10 runs

# after
❯ hyperfine "neut clean && neut build test"
Benchmark 1: neut clean && neut build test
  Time (mean ± σ):     588.3 ms ±   6.8 ms    [User: 1405.3 ms, System: 548.2 ms]
  Range (min … max):   580.5 ms … 598.6 ms    10 runs
```